Dev 3.1.0 haas (#1285)
* add haascv module & haascv demo


* upate haascv static lib & add mpy demo

+++ components/service/ai/uai/function/include/uai_pool.h
@@ -13,9 +13,9 @@ extern "C" {
 #endif
 
 #ifdef UAI_ODLA_SUPPORT
-void uai_pool_max(uai_tensor_s *input, uai_tensor_s *weight, uint16_t *strides, const unsigned *paddings_front,
+void uai_pool_max(uai_tensor_s *input, uai_tensor_s *weight, uint32_t *strides, const unsigned *paddings_front,
                   const unsigned *paddings_back, uai_tensor_s *output);
-void uai_pool_ave(uai_tensor_s *input, uai_tensor_s *weight, uint16_t *strides, const unsigned *paddings_front,
+void uai_pool_ave(uai_tensor_s *input, uai_tensor_s *weight, uint32_t *strides, const unsigned *paddings_front,
                   const unsigned *paddings_back, uai_tensor_s *output);
 #else
 void uai_pool_max(uai_input_s *input, uai_weight_s *filter, uint16_t *strides, uai_pad_type_e pad_type,
